Prevent duplicate initialization of DataTables on course reserves
