+++ app/javascript/application.js
@@ -25,7 +25,6 @@ import "./alternate_catalog";
 import "./analytics";
 import "./article";
 import "./async_collection_members";
-import "./course_reserves";
 import "./eds_range_limit";
 import "./embedded-call-number-browse";
 import "./exhibitPanel";
