+++ app/views/course_reserves/index.html.erb
@@ -11,7 +11,7 @@
   </div>
 </div>
 
-<table id="course-reserves-browse" class="table table-striped">
+<table id="course-reserves-browse" class="table table-striped" data-controller="course-reserves">
   <thead>
     <tr>
       <th>
